# faithgpt-platform/api/openapi.yaml
openapi: 3.0.3
info:
  title: FaithGPT API
  version: "1.0.0"
  description: >
    Canonical REST API contract for FaithGPT™ ("Transforming Scripture into Daily Living").
    Conventions, envelopes, pagination, streaming, and rate-limiting are documented narratively
    in /faithgpt-platform/docs/09-api-architecture.md — this document is the reviewable
    machine-readable snapshot. Schema names mirror /faithgpt-platform/schema/schema.prisma
    verbatim (see /faithgpt-platform/docs/00-shared-reference.md §13).
  contact:
    name: FaithGPT Engineering
    url: https://faithgpt.app
servers:
  - url: https://api.faithgpt.app/api/v1
    description: Production
  - url: https://staging-api.faithgpt.app/api/v1
    description: Staging
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []

# =====================================================================================
# PATHS
# =====================================================================================
paths:
...
  /organizations:
    get:
      tags: [Organizations]
      summary: List organizations the authenticated user belongs to
      responses:
        "200":
          description: Organization memberships
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Organization'
    post:
      tags: [Organizations]
      summary: Create a new Ministry organization (initiates MINISTRY subscription setup)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                type: { type: string, example: "CHURCH" }
                billingEmail: { type: string, format: email, nullable: true }
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOrganization'
  /organizations/{id}:
    get:
      tags: [Organizations]
      summary: Get organization detail
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOrganization'
        "404":
          $ref: '#/components/responses/NotFound'
        "403":
          $ref: '#/components/responses/Forbidden'
    patch:
      tags: [Organizations]
      summary: Update organization (name, seatLimit, billingEmail) — ADMIN role in org
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                seatLimit: { type: integer }
                billingEmail: { type: string, format: email, nullable: true }
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOrganization'
  /organizations/{id}/members:
    get:
      tags: [Organizations]
      summary: List organization members
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrganizationMember'
                  meta:
                    $ref: '#/components/schemas/OffsetPaginationMeta'
    post:
      tags: [Organizations]
      summary: Invite/add a member to the organization
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email, roleInOrg]
              properties:
                email: { type: string, format: email }
                roleInOrg: { type: string, enum: [ADMIN, STAFF] }
      responses:
        "201":
          description: Member added/invited
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/OrganizationMember'

  # ---------------------------------------------------------------------------------
  # SUBSCRIPTIONS
  # ---------------------------------------------------------------------------------
...
  /journal-entries:
    get:
      tags: [Journal]
      summary: List journal entries
      parameters:
        - name: type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/JournalEntryType'
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        "200":
          description: Journal entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/JournalEntry'
                  meta:
                    $ref: '#/components/schemas/CursorPaginationMeta'
    post:
      tags: [Journal]
      summary: Create a journal entry
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [type, content]
              properties:
                type:
                  $ref: '#/components/schemas/JournalEntryType'
                title: { type: string, nullable: true }
                content: { type: string }
                tags:
                  type: array
                  items: { type: string }
                relatedDevotionId: { type: string, nullable: true }
                relatedPrayerId: { type: string, nullable: true }
      responses:
        "201":
          description: Entry created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/JournalEntry'
  /journal-entries/{id}:
    get:
      tags: [Journal]
      summary: Get a journal entry
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "200":
          description: Journal entry
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/JournalEntry'
    patch:
      tags: [Journal]
      summary: Update a journal entry
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string, nullable: true }
                content: { type: string }
                tags:
                  type: array
                  items: { type: string }
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/JournalEntry'
    delete:
      tags: [Journal]
      summary: Delete a journal entry
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "204":
          description: Deleted
  /prayer-requests:
    get:
      tags: [Journal]
      summary: List prayer requests (own, or group/public per visibility)
      parameters:
        - name: visibility
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/PrayerRequestVisibility'
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        "200":
          description: Prayer requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PrayerRequest'
                  meta:
                    $ref: '#/components/schemas/CursorPaginationMeta'
    post:
      tags: [Journal]
      summary: Create a prayer request
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
                isAnonymous: { type: boolean, default: false }
                visibility:
                  $ref: '#/components/schemas/PrayerRequestVisibility'
                groupId: { type: string, nullable: true }
      responses:
        "201":
          description: Prayer request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/PrayerRequest'
  /prayer-requests/{id}/answer:
    post:
      tags: [Journal]
      summary: Mark a prayer request as answered
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                answeredNote: { type: string, nullable: true }
      responses:
        "200":
          description: Updated prayer request
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/PrayerRequest'

  # ---------------------------------------------------------------------------------
  # GROWTH
  # ---------------------------------------------------------------------------------
...
  /notifications:
    get:
      tags: [Notifications]
      summary: List the authenticated user's notifications
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
                  meta:
                    $ref: '#/components/schemas/CursorPaginationMeta'
  /notifications/{id}/read:
    post:
      tags: [Notifications]
      summary: Mark a notification as read
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "204":
          description: Marked read
  /notifications/preferences:
    get:
      tags: [Notifications]
      summary: Get notification preferences (by NotificationType)
      responses:
        "200":
          description: Preferences
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    additionalProperties: { type: boolean }
                    example:
                      DAILY_DEVOTION: true
                      STREAK_REMINDER: true
                      PRAYER_ANSWERED: true
                      GROUP_ACTIVITY: true
                      COMMUNITY_INTERACTION: false
                      SUBSCRIPTION: true
                      SYSTEM: true
    patch:
      tags: [Notifications]
      summary: Update notification preferences (transactional types always-on, non-editable)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: { type: boolean }
      responses:
        "200":
          description: Updated preferences

  # ---------------------------------------------------------------------------------
  # ADMIN (subset — see §17 Admin Portal Design for full UI)
  # ---------------------------------------------------------------------------------
...
  /admin/organizations/{id}:
    patch:
      tags: [Admin]
      summary: Update an organization's seats/billing (ADMIN, SUPER_ADMIN)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                seatLimit: { type: integer }
                billingEmail: { type: string, format: email, nullable: true }
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Organization'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: Opaque cuid()
    CursorParam:
      name: cursor
      in: query
      required: false
      schema: { type: string }
    LimitParam:
      name: limit
      in: query
      required: false
      schema: { type: integer, default: 20, maximum: 100 }
    PageParam:
      name: page
      in: query
      required: false
      schema: { type: integer, default: 1 }
    PageSizeParam:
      name: pageSize
      in: query
      required: false
      schema: { type: integer, default: 20, maximum: 100 }
  responses:
...
    Forbidden:
      description: Authenticated but not authorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: "VALIDATION_ERROR" }
            message: { type: string }
            details:
              type: object
              nullable: true
              additionalProperties: true
    CursorPaginationMeta:
      type: object
      properties:
        pagination:
          type: object
          properties:
            nextCursor: { type: string, nullable: true }
            hasMore: { type: boolean }
    OffsetPaginationMeta:
      type: object
      properties:
        pagination:
          type: object
          properties:
            page: { type: integer }
            pageSize: { type: integer }
            total: { type: integer }
            totalPages: { type: integer }
...
    EnvelopeOrganization:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Organization'

    # --- Enums (mirror schema.prisma verbatim, §00 §13) ---------------------
    UserRole:
      type: string
      enum: [USER, GROUP_LEADER, MINISTRY_ADMIN, MODERATOR, CONTENT_EDITOR, SUPPORT, ADMIN, SUPER_ADMIN]
    SubscriptionTier:
      type: string
      enum: [FREE, PLUS, PREMIUM, MINISTRY, LIFETIME]
...
    JournalEntryType:
      type: string
      enum: [NOTE, DEVOTION, PRAYER, TESTIMONY, INSIGHT, ANSWERED_PRAYER]
    PrayerRequestVisibility:
      type: string
      enum: [PRIVATE, GROUP, PUBLIC]
...
    NotificationType:
      type: string
      enum: [DAILY_DEVOTION, STREAK_REMINDER, PRAYER_ANSWERED, GROUP_ACTIVITY, COMMUNITY_INTERACTION, SUBSCRIPTION, SYSTEM]
...
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email, nullable: true }
        phone: { type: string, nullable: true }
        name: { type: string }
        avatarUrl: { type: string, nullable: true }
        role:
          $ref: '#/components/schemas/UserRole'
        tier:
          $ref: '#/components/schemas/SubscriptionTier'
        denominationLens: { type: string, nullable: true, example: "NEUTRAL" }
        locale: { type: string, example: "en" }
        timezone: { type: string, example: "UTC" }
        defaultBibleVersionId: { type: string, nullable: true }
        emailVerifiedAt: { type: boolean }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    Organization:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string, example: "CHURCH" }
        seatLimit: { type: integer }
        billingEmail: { type: string, format: email, nullable: true }
        createdAt: { type: string, format: date-time }
    OrganizationMember:
      type: object
      properties:
        id: { type: string }
        organizationId: { type: string }
        userId: { type: string }
        roleInOrg: { type: string, enum: [ADMIN, STAFF] }
        joinedAt: { type: string, format: date-time }
        user:
          $ref: '#/components/schemas/User'
...
    JournalEntry:
      type: object
      properties:
        id: { type: string }
        type:
          $ref: '#/components/schemas/JournalEntryType'
        title: { type: string, nullable: true }
        content: { type: string }
        tags:
          type: array
          items: { type: string }
        relatedDevotionId: { type: string, nullable: true }
        relatedPrayerId: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PrayerRequest:
      type: object
      properties:
        id: { type: string }
        content: { type: string }
        isAnonymous: { type: boolean }
        isAnswered: { type: boolean }
        answeredNote: { type: string, nullable: true }
        answeredAt: { type: string, format: date-time, nullable: true }
        visibility:
          $ref: '#/components/schemas/PrayerRequestVisibility'
        groupId: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }

    # --- Growth ---------------------------------------------------------
...
    Notification:
      type: object
      properties:
        id: { type: string }
        type:
          $ref: '#/components/schemas/NotificationType'
        title: { type: string }
        body: { type: string }
        deepLink: { type: string, nullable: true, example: "faithgpt://devotion/clx1devotion0001" }
        readAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }

    # --- Admin ---------------------------------------------------------
...